refactor: server error maintenance
