+++ project/src/components/oops/oops.module.sass
@@ -5,36 +5,39 @@
   padding: 0
 
 /* I. Containers */
-figure
+.figureoops
   font-size: 6px
   margin: auto
   padding: 4em 0
-  width: 40em
+  width: 37em
 
-figcaption
+.figcaptionoops
   color: #fff
   display: flex
   align-content: space-between
   flex-wrap: wrap
-  height: 17em
+  height: 8em
+  cursor: pointer
 
-figcaption span:before
+.figcaptionoops span:before
   content: ""
   display: block
   width: 1em
   height: 1em
   transform: translate(-1em,-1em)
 
-figcaption span
+.figcaptionoops span
   display: inline-block
   margin: 0 2em
   width: 4em
   height: 7em
 
-.container
+.figcaptionoops span.wow
+  width: 1em
+
+.containeroops
   background: currentColor
   height: 100vh
-  padding-left: 8em
   padding-top: 42vh
 
 .o:before
