+++ project/src/pages/main/main.tsx
@@ -9,13 +9,13 @@ import Locations from '../../components/locations/locations';
 import Sort from '../../components/sort/sort';
 import CityMap from '../../components/city-map/city-map';
 import LoadingSpinner from '../../components/loading-spinner/loading-spinner';
-import Oops from '../oops/oops';
 
 import { CurrentSortCallback } from '../../utils/sort-offers';
 import { Cities, FetchStatus, SortType } from '../../const';
 import { getCurrentCity, getCurrentSort, getOffers, getOffersLoadingStatus } from '../../store/app/selectors';
 import { changeCurrentCity } from '../../store/app/actions';
 import { fetchOffers } from '../../store/app/api-actions';
+import Oops from '../../components/oops/oops';
 
 export default function Main(): JSX.Element {
   const offers = useAppSelector(getOffers);
