+++ project/src/pages/not-found/not-found.module.sass
@@ -5,33 +5,33 @@
   padding: 0
 
 /* I. Containers */
-figure
+.figure404
   font-size: 6px
   margin: auto
   padding: 4em 0
   width: 64em
 
-figcaption
+.figcaption404
   color: #fff
   display: flex
   align-content: space-between
   flex-wrap: wrap
   height: 17em
 
-figcaption span:before
+.figcaption404 span:before
   content: ""
   display: block
   width: 1em
   height: 1em
   transform: translate(-1em,-1em)
 
-figcaption span
+.figcaption404 span
   display: inline-block
   margin: 0 2em
   width: 4em
   height: 6em
 
-.container
+.container404
   background: currentColor
   height: 100vh
   padding-top: 35vh
